requirements-docker.txt
# Docker-optimized requirements for Modern Hosting Panel
# These versions are tested and compatible with Python 3.11+ and Docker

# Web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP and forms
python-multipart>=0.0.6
requests>=2.31.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1
jinja2>=3.1.2
psutil>=5.9.6

# Docker integration
docker>=6.1.3

# SSL and certificates
certbot>=2.7.4
dnspython>=2.4.2

# Caching and background tasks
redis>=5.0.1
celery>=5.3.4
